fix script, for delete cache directory

+++ scripts/clearCache.js
@@ -2,15 +2,19 @@ const path = require("path");
 const fs = require("fs");
 
 function clearCache() {
-	fs.rm(
-		path.join(__dirname, "..", "node_modules", ".cache"),
-		{ recursive: true },
-		(err) => {
-			if (err) {
-				throw err;
-			}
-		},
-	);
+	const pathCache = path.join(__dirname, "..", "node_modules", ".cache");
+
+	if (pathCache) {
+		fs.rm(
+			pathCache,
+			{ recursive: true },
+			(err) => {
+				if (err) {
+					throw err;
+				}
+			},
+		);
+	}
 }
 
 clearCache();
